test: split out getNestedCodes
